Move all the NumPy functionality into a numpy submodule
There will also be a 'cupy' submodule. Common code will be factored out into
the 'common' submodule.
