+++ numpy_array_api_compat/__init__.py
@@ -40,21 +40,3 @@ Known differences from the Array API spec:
 - NumPy functions which are not wrapped may not use positional-only arguments.
 
 """
-
-from numpy import *
-
-# These imports may overwrite names from the import * above.
-from ._aliases import *
-
-# Don't know why, but we have to do an absolute import to import linalg. If we
-# instead do
-#
-# from . import linalg
-#
-# It doesn't overwrite np.linalg from above. The import is generated
-# dynamically so that the library can be vendored.
-__import__(__package__ + '.linalg')
-
-from .linalg import matrix_transpose, vecdot
-
-from ._helpers import *
